RHOL-504: Move StreamT to separate file
