+++ rholang/src/main/scala/coop/rchain/rholang/interpreter/matcher/SpatialMatcher.scala
@@ -14,8 +14,8 @@ import coop.rchain.models.Expr.ExprInstance._
 import coop.rchain.models.Var.VarInstance.{BoundVar, FreeVar, Wildcard}
 import coop.rchain.models._
 import coop.rchain.models.rholang.implicits.{VectorPar, _}
-import coop.rchain.rholang.interpreter.StreamT
-import coop.rchain.rholang.interpreter.StreamT._
+import coop.rchain.rholang.interpreter.matcher.StreamT
+import coop.rchain.rholang.interpreter.matcher.StreamT._
 import coop.rchain.rholang.interpreter.accounting.CostAccount
 import coop.rchain.rholang.interpreter.matcher.SpatialMatcher._
 import OptionalFreeMapWithCost._
